Add files via upload
mapsHome now has a working model of two different google maps maintained in the same location. Upon pressing "Switch", the hidden maps center and zoom are aligned with the visible one and they are switched. For now they are both heatmaps, but this will be used to switch between a heatmap and a map with clustered markers.