Add files via upload
mapsHome.html now includes two different map types to switch between; a heatmap and a clustered markers map. The latter relies on markerClusterer.js as a class definition, as well as five different images to act as the cluster image (will be uploaded later). The other html files have had reference changes to the locations of their images.