Add files via upload
reportsHome now connects to Apps Script after prompting user for google sign in; it's problematic, but it connects. chartsHome now has chart.js. The only thing left the do on charts is allow customization of the data displayed, but the same goes for all three pages. I'm uploading mapsHome stuff because I probably changed some of it too, but I cant remember what.